# storage/docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OmPay API",
    "description": "API de paiement mobile OmPay avec authentification OTP",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://ompay-api-ou4t.onrender.com",
      "description": "Serveur de production Render"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Inscription d'un nouvel utilisateur",
        "description": "Crée un compte utilisateur et envoie automatiquement un code OTP par SMS",
        "tags": ["Authentification"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nom", "telephone", "pin"],
                "properties": {
                  "nom": {"type": "string", "example": "Jean Dupont"},
                  "telephone": {"type": "string", "example": "+221+221771234567"},
                  "pin": {"type": "string", "example": "1234"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur créé avec succès",
            "content": {
              "application/json": {
                "example": {
                  "succes": true,
                  "message": "Utilisateur créé avec succès. Vérifiez votre téléphone pour le code OTP.",
                  "donnees": {
                    "utilisateur": {
                      "id": 1,
                      "uuid": "550e8400-e29b-41d4-a716-446655440000",
                      "nom": "Jean Dupont",
                      "telephone": "+221771234567"
                    },
                    "message_complementaire": "Un code OTP a été envoyé à votre téléphone pour finaliser l'inscription"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Données invalides",
            "content": {
              "application/json": {
                "example": {
                  "succes": false,
                  "message": "Les données fournies ne sont pas valides",
                  "erreurs": {
                    "telephone": ["Ce numéro de téléphone est déjà utilisé"]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Connexion utilisateur",
        "description": "Authentifie l'utilisateur et envoie un code OTP par SMS",
        "tags": ["Authentification"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["telephone", "pin"],
                "properties": {
                  "telephone": {"type": "string", "example": "+221+221771234567"},
                  "pin": {"type": "string", "example": "1234"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code OTP envoyé",
            "content": {
              "application/json": {
                "example": {
                  "succes": true,
                  "message": "Veuillez vérifier votre téléphone avec le code OTP",
                  "donnees": {
                    "utilisateur": {
                      "id": 1,
                      "uuid": "550e8400-e29b-41d4-a716-446655440000",
                      "nom": "Jean Dupont",
                      "telephone": "+221771234567"
                    },
                    "message_complementaire": "Un code OTP a été envoyé à votre numéro de téléphone"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur non trouvé",
            "content": {
              "application/json": {
                "example": {
                  "succes": false,
                  "message": "Numéro de téléphone non enregistré"
                }
              }
            }
          },
          "401": {
            "description": "PIN incorrect",
            "content": {
              "application/json": {
                "example": {
                  "succes": false,
                  "message": "PIN incorrect"
                }
              }
            }
          }
        }
      }
    },
...
  }
}